reorganize directories structure
