+++ .gitignore
@@ -1,3 +1,4 @@
-_build
-_install
+_build/
+_install/
+share/
 .vscode
