+++ autonomy/src/control/position_controller.cpp
@@ -1,4 +1,4 @@
-#include "control/position_controller.hpp"
+#include "autonomy/control/position_controller.hpp"
 #include <math.h>
 
 PosCtrl::PosCtrl(double len, double gain_x, double gain_y) {
